extended pymol with logging
