+++ ciftools/Residue.py
